Fix delete() function on RedisTreeEngine
The delete() function on RedisTreeEngine is finding all keys first (with
the KEYS command) and then delete them with one command (with the DEL
command). But if no keys was found the DEL command was being called with
an empty array, causing an error.
This issue was not caught by the unit test because the mock class allows
passing an empty array the the del() function, unlike the real Redis
driver.