Fix FileTreeEngine::clear()
Function was originally overwritten but it was causing all files in the
cache dir to be deleted. The original version (from FileEngine)
correctly only delete file that start with the prefix set for that
cache.
Add entry to bootstarp to make sure test/tmp is created even though the
unit tests do not delete all its content anyore (i.e. the .gitignore
file is preserved).