Support multiple Fallback caches
The Fallback engine configures 2 caches internally. The 2 configurations
for these caches where initally called 'primary' and 'secondary'. But
this creates a conflict if multiple Fallback caches are creates.
Now require a 'name' setting for the Fallback engine and use it to
create unique config names for the 2 internal engines.
Also make sure to call parent:init() as this does some internal work
like converting duration when it's set with '+x days' for instance.